refactor(onboarding): split step components
