+++ src/components/welcome/onboarding/steps/CompletionStep/index.tsx
@@ -6,6 +6,7 @@ import { CheckCircle2, Sparkles, Star, Folder, TrendingUp } from 'lucide-react';
 import { getMessage } from '@/core/utils/i18n';
 import { trackEvent, EVENTS } from '@/utils/amplitude';
 import { AIToolGrid } from '@/components/welcome/AIToolGrid';
+import { CelebrationStars } from './CelebrationStars';
 
 interface FolderDetail {
   id: number;
@@ -189,39 +190,5 @@ export const CompletionStep: React.FC<CompletionStepProps> = ({
         </div>
       </motion.div>
     </motion.div>
-  );
-};
-
-// Celebration animation component
-const CelebrationStars = () => {
-  return (
-    <>
-      {[...Array(8)].map((_, i) => (
-        <motion.div
-          key={i}
-          className="jd-absolute"
-          style={{
-            top: '50%',
-            left: '50%',
-            x: `-50%`,
-            y: `-50%`,
-          }}
-          initial={{ scale: 0, opacity: 0 }}
-          animate={{ 
-            scale: [0, 1.5, 0],
-            opacity: [0, 1, 0],
-            x: [0, (i % 2 ? 60 : -60) * Math.sin(i * 45 * Math.PI / 180)],
-            y: [0, (i % 2 ? 60 : -60) * Math.cos(i * 45 * Math.PI / 180)], 
-          }}
-          transition={{ 
-            duration: 1,
-            delay: 0.3 + (i * 0.05),
-            ease: "easeOut" 
-          }}
-        >
-          <Star className="jd-h-4 jd-w-4 jd-text-yellow-300" />
-        </motion.div>
-      ))}
-    </>
   );
 };
