+++ src/components/welcome/onboarding/steps/ReferralStep/index.tsx
@@ -5,6 +5,7 @@ import { Users } from 'lucide-react';
 import { getMessage } from '@/core/utils/i18n';
 import { trackEvent, EVENTS } from '@/utils/amplitude';
 import { OnboardingData } from '../OnboardingFlow';
+import { REFERRAL_SOURCES } from './constants';
 
 // Components
 import { RadioGroup } from '@/components/ui/radio-group';
@@ -12,46 +13,6 @@ import { OnboardingOption } from '@/components/welcome/onboarding/OnboardingOpti
 import { OnboardingActions } from '@/components/welcome/onboarding/OnboardingActions';
 import { OtherOptionInput } from '@/components/welcome/onboarding/OtherOptionInput';
 
-// Referral sources
-const REFERRAL_SOURCES = [
-  {
-    value: 'search',
-    label: getMessage(
-      'referralSourceSearch',
-      undefined,
-      'Search Engine (Google, Bing, etc.)'
-    )
-  },
-  {
-    value: 'social_media',
-    label: getMessage('referralSourceSocialMedia', undefined, 'Social Media')
-  },
-  {
-    value: 'friend',
-    label: getMessage('referralSourceFriend', undefined, 'Friend or Colleague')
-  },
-  {
-    value: 'blog',
-    label: getMessage('referralSourceBlog', undefined, 'Blog or Article')
-  },
-  {
-    value: 'youtube',
-    label: getMessage('referralSourceYouTube', undefined, 'YouTube')
-  },
-  {
-    value: 'podcast',
-    label: getMessage('referralSourcePodcast', undefined, 'Podcast')
-  },
-  {
-    value: 'ad',
-    label: getMessage('referralSourceAd', undefined, 'Advertisement')
-  },
-  {
-    value: 'store',
-    label: getMessage('referralSourceStore', undefined, 'Chrome Web Store')
-  },
-  { value: 'other', label: getMessage('referralSourceOther', undefined, 'Other') }
-];
 
 interface ReferralStepProps {
   initialData: OnboardingData;
